First OpenGTO model
This is just a preflop model. Please view the README.md for more details. This codebase is already very messy with old files in the poker_gto folder. Please disregard any old files. These will be removed in the future.

Currently it doesn't take into account blinds correctly and how it would impact the game, furthermore it doesn't know that standing is a thing yet which is a bit cursed. Also the training script is broken so it only trains up to 50,000 examples and then truncates itself.